requirements.txt
flask==2.3.2
flask-cors==4.0.0
numpy==2.0.2
ultralytics==8.3.39
Pillow==10.1.0
opencv-python==4.10.0.84